Added relevant Hackney styles.
Added them to global CSS as they give a binch of issues when I add them to SCSS file. The issues are related to the syntax of MAX and CALC functions. I could add these style to index.module.css and import that, however I'm going to leave it for now and see how development goes.