Prepare application for 100% height scaling.
Only the html element can be scaled in %'ages. All other children elements only get scaled to the % height of parent container. So effectively, if you want to scale body tag to occupy 100% of available browser window height with height:100%, it will only be able to scale up to the full size of its immediate parent, which would be html element. Html by default, at least here is not 100% window height, hence the body tag won
't be either. So what I did here was set html to 100% as its immediate parent is the window, meaning html will occupy 100% of (parent) window height. Then I set the body to 100%, which will grow to 100% height of (parent) html. Lastly, I set #__next to 100%, which will grow to 100% height of (parent) body tag. Hence, this will permit me to grow any of my markup divs to 100% window height whenever I need to - and I will need for the sidebar. Tldr; When using percentages, there's a chain of command that carries over the full window height: window --> html --> body --> #__next --> any of my divs on the home page or layout component.