Added main data model classes.
At this point it's hard to say whether they're going to changes. Ideally, I want every single field listed here. Practically, they're only as useful as supplied data. Supplying the data is a challenge that will require a lot of work post MVP, hence initially, not everything is going to be used.